@@ -21,35 +21,35 @@
 	"license": "MIT",
 	"main": "dist/main/index.js",
 	"dependencies": {
-		"electron-store": "^8.0.1",
-		"electron-updater": "^4.6.5",
-		"nanoid": "^3.3.1",
-		"prettier": "^2.5.1"
+		"electron-store": "^8.1.0",
+		"electron-updater": "^5.3.0",
+		"nanoid": "^3.3.4",
+		"prettier": "^2.7.1"
 	},
 	"devDependencies": {
-		"@fontsource/roboto": "^4.5.3",
-		"@types/mocha": "^9.1.0",
-		"@types/prettier": "^2.4.4",
-		"@types/react": "^17.0.33",
-		"@types/react-dom": "^17.0.10",
-		"@vitejs/plugin-react": "^1.0.7",
-		"chart.js": "^3.7.1",
-		"concurrently": "^7.0.0",
-		"electron": "17.0.1",
-		"electron-builder": "^22.14.13",
+		"@fontsource/roboto": "^4.5.8",
+		"@types/mocha": "^10.0.0",
+		"@types/prettier": "^2.7.1",
+		"@types/react": "^18.0.25",
+		"@types/react-dom": "^18.0.9",
+		"@vitejs/plugin-react": "^2.2.0",
+		"chart.js": "^4.0.1",
+		"concurrently": "^7.5.0",
+		"electron": "21.3.0",
+		"electron-builder": "^23.6.0",
 		"electron-devtools-installer": "^3.2.0",
-		"expect": "^27.5.1",
-		"fuzzysort": "^1.2.1",
-		"mocha": "^9.2.1",
-		"react": "^17.0.2",
-		"react-chartjs-2": "^4.0.1",
-		"react-dom": "^17.0.2",
-		"react-icons": "^4.3.1",
-		"react-router-dom": "^6.2.2",
-		"recoil": "^0.6.1",
-		"ts-node": "^10.7.0",
-		"typescript": "^4.5.4",
-		"vite": "^2.8.0",
-		"vite-plugin-resolve": "^1.8.0"
+		"expect": "^29.3.1",
+		"fuzzysort": "^2.0.3",
+		"mocha": "^10.1.0",
+		"react": "^18.2.0",
+		"react-chartjs-2": "^5.0.1",
+		"react-dom": "^18.2.0",
+		"react-icons": "^4.6.0",
+		"react-router-dom": "^6.4.3",
+		"recoil": "^0.7.6",
+		"ts-node": "^10.9.1",
+		"typescript": "^4.9.3",
+		"vite": "^3.2.4",
+		"vite-plugin-resolve": "^2.3.0"
 	}
 }
