@@ -3,7 +3,9 @@ import React from 'react'
 import Navigation from '../Navigation'
 import styles from './Layout.module.css'
 
-const Layout: React.FC = ({ children }) => (
+type Props = { children?: React.ReactNode }
+
+const Layout: React.FC<Props> = ({ children }) => (
 	<div className={styles.root}>
 		<React.Suspense fallback={<div>Loading...</div>}>
 			<Navigation />
